requirements.txt
streamlit
pathlib
google-generativeai
IPython
python-dotenv
fitz
PyMuPDF

matplotlib
yfinance
